Change async_stdin to return a result
If called often, this function periodically fails. (It can be the case that the stream is unavailable). This changes the async_stdin API to return a result instead of using unwrap. Luckily, other internal functions which use async_stdin were already returning results.

Still, this is a breaking change to the external API